+++ example/shape/pipe.py
@@ -1,7 +1,9 @@
 import cadquery as cq
-from cadqueryhelper import shape
+from cadqueryhelper.shape import pipe
 
-result = shape.pipe()
+test_shape = cq.Workplane('XY').rect(4,8)
+pst = pts = [(0,0), (20,-20), (50,-20), (50,-30)]
+result = pipe(test_shape, pts)
 
 #show_object(result)
 cq.exporters.export(result,'stl/shape_pipe.stl')