+++ src/cadqueryhelper/shape/__init__.py
@@ -23,3 +23,4 @@ from .pinwheel import make_circular_points, interweave_lists, pinwheel
 from .trapezoid import trapezoid
 from .backdrop import backdrop
 from .jersey_barrier import jersey_barrier
+from .vase import vase